Add ex02: Implement Runtime Type Identification
This commit adds the necessary files to implement Runtime Type
Identification (RTTI) in the ex02 directory. It includes the definition
of base class `Base`, and derived classes `A`, `B`, and `C` along with a
`main.cpp` file to test RTTI functionality using `dynamic_cast`. A
`Makefile` is also included to build the program.